+++ lib/msix.dart
@@ -1,5 +1,4 @@
 import 'dart:io';
-
 import 'package:cli_util/cli_logging.dart';
 import 'package:get_it/get_it.dart';
 import 'src/app_installer.dart';
@@ -10,7 +9,7 @@ import 'src/makepri.dart';
 import 'src/appx_manifest.dart';
 import 'src/makeappx.dart';
 import 'src/sign_tool.dart';
-import 'src/extensions.dart';
+import 'src/method_extensions.dart';
 
 /// Main class that handles all the msix package functionality
 class Msix {
