+++ lib/src/makepri.dart
@@ -2,7 +2,7 @@ import 'dart:io';
 import 'package:cli_util/cli_logging.dart';
 import 'package:get_it/get_it.dart';
 import 'configuration.dart';
-import 'extensions.dart';
+import 'method_extensions.dart';
 
 /// Use the makepri.exe tool to generate package resource indexing files
 class MakePri {
@@ -16,16 +16,16 @@ class MakePri {
     String makePriPath =
         '${_config.msixToolkitPath}/Redist.${_config.architecture}/makepri.exe';
 
-    ProcessResult makePriConfigProcess = await Process.run(makePriPath, [
+    // ignore: avoid_single_cascade_in_expression_statements
+    await Process.run(makePriPath, [
       'createconfig',
       '/cf',
       '$buildPath\\priconfig.xml',
       '/dq',
       'en-US',
       '/o'
-    ]);
-
-    makePriConfigProcess.exitOnError();
+    ])
+      ..exitOnError();
 
     ProcessResult makePriProcess = await Process.run(makePriPath, [
       'new',
