+++ lib/src/method_extensions.dart
