+++ lib/src/windows_build.dart
@@ -1,8 +1,7 @@
 import 'dart:io';
 import 'package:cli_util/cli_logging.dart';
 import 'package:get_it/get_it.dart';
-import 'package:msix/src/extensions.dart';
-
+import 'method_extensions.dart';
 import 'configuration.dart';
 
 const runnerRcPath = 'windows/runner/Runner.rc';
@@ -21,10 +20,9 @@ class WindowsBuild {
     Progress loggerProgress = _logger
         .progress('running "flutter ${buildWindowsArguments.join(' ')}"');
 
-    ProcessResult windowsBuildProcess =
-        await Process.run('flutter', buildWindowsArguments, runInShell: true);
-
-    windowsBuildProcess.exitOnError();
+    // ignore: avoid_single_cascade_in_expression_statements
+    await Process.run('flutter', buildWindowsArguments, runInShell: true)
+      ..exitOnError();
 
     loggerProgress.finish(showTiming: true);
   }
