+++ test/configuration_test.dart
@@ -26,16 +26,13 @@ msix_config:
     GetIt.I.registerSingleton<Configuration>(config);
 
     await Directory('$tempFolderPath/').create(recursive: true);
-    await Future.delayed(const Duration(milliseconds: 150));
   });
 
   tearDown(() async {
     GetIt.I.reset();
 
     if (await Directory('$tempFolderPath/').exists()) {
-      await Future.delayed(const Duration(milliseconds: 150));
       await Directory('$tempFolderPath/').delete(recursive: true);
-      await Future.delayed(const Duration(milliseconds: 150));
     }
   });
